requirements.txt
django
djangorestframework
django-query-counter